requirements.txt
Django==1.8.2
psycopg2==2.6
DjPyBinder==2.0.1
PyYAML==3.11
